Consolidate DDL tests
